refactor: change sql files
1. change id type to SERIAL

+++ db/schema/slack_users.sql
@@ -1,5 +1,5 @@
 CREATE TABLE slack_users (
-    slack_user_id BIGINT NOT NULL PRIMARY KEY,
+    slack_user_id SERIAL NOT NULL PRIMARY KEY,
     id VARCHAR(255) NOT NULL,
     name VARCHAR(255) NOT NULL,
     real_name VARCHAR(255) null,
